+++ lib/balance_lib.ex
@@ -2,19 +2,52 @@ defmodule Teiserver.Battle.BalanceLib do
   @moduledoc """
   A set of functions related to balance, if you are looking to see how balance is implemented this is the place. Ratings are calculated via Teiserver.Game.MatchRatingLib and are used here. Please note ratings and balance are two very different things and complaints about imbalanced games need to be correct in addressing balance vs ratings.
   """
+  alias Teiserver.{Account, Config}
   alias Teiserver.Data.Types, as: T
-  alias Teiserver.Account
+  alias Teiserver.Battle.Balance.BalanceTypes, as: BT
   alias Teiserver.Game.MatchRatingLib
-  alias Teiserver.Battle.LoserPicksAlgorithm
-  alias Teiserver.Battle.CheekySwitcherAlgorithm
-  alias Teiserver.Battle.CheekySwitcherSmartAlgorithm
-  alias Teiserver.Battle.BruteForceAlgorithm
-  import Central.Helpers.NumberHelper, only: [int_parse: 1]
-  import Teiserver.Battle.BalanceUtil
+  import Teiserver.Helper.NumberHelper, only: [int_parse: 1, round: 2]
 
-  @type player_group :: BalanceUtils.player_group
-  @type rating_value :: BalanceUtils.rating_value
+  # These are default values and can be overridden as part of the call to create_balance()
 
+  # Upper boundary is how far above the group value the members can be, lower is how far below it
+  # these values are in general used for group pairing where we look at making temporary groups on
+  # each team to make the battle fair
+  @rating_lower_boundary 3
+  @rating_upper_boundary 5
+
+  @mean_diff_max 5
+  @stddev_diff_max 3
+
+  # Fuzz multiplier is used by the BalanceServer to prevent two games being completely identical
+  # teams. It is defaulted here as the server uses this library to get defaults
+  @fuzz_multiplier 0.5
+
+  # When set to true, if there are any teams with 0 points (first pick) it randomises
+  # which one will get to pick first
+  @shuffle_first_pick true
+
+  @spec defaults() :: map()
+  def defaults() do
+    %{
+      max_deviation: Config.get_site_config_cache("teiserver.Max deviation"),
+      rating_lower_boundary: @rating_lower_boundary,
+      rating_upper_boundary: @rating_upper_boundary,
+      mean_diff_max: @mean_diff_max,
+      stddev_diff_max: @stddev_diff_max,
+      fuzz_multiplier: @fuzz_multiplier,
+      shuffle_first_pick: @shuffle_first_pick
+    }
+  end
+
+  @spec algorithm_modules() :: %{String.t() => module}
+  def algorithm_modules() do
+    %{
+      "loser_picks" => Teiserver.Battle.Balance.LoserPicks,
+      "force_party" => Teiserver.Battle.Balance.ForceParty,
+      "cheeky_switcher_smart" => Teiserver.Battle.Balance.CheekySwitcherSmart
+    }
+  end
 
   @doc """
   groups is a list of maps of %{userid => rating_value}
@@ -28,7 +61,7 @@ defmodule Teiserver.Battle.BalanceLib do
   team_groups: map of team_id => list of expanded_groups
 
   Options are:
-    mode
+    algorithm: String name of the algorithm
 
     rating_lower_boundary: the amount of rating points to search below a party
     rating_upper_boundary: the amount of rating points to search above a party
@@ -36,9 +69,23 @@ defmodule Teiserver.Battle.BalanceLib do
     mean_diff_max: the maximum difference in mean between the party and paired parties
     stddev_diff_max: the maximum difference in stddev between the party and paired parties
   """
-  @spec create_balance([player_group()], non_neg_integer()) :: map()
-  @spec create_balance([player_group()], non_neg_integer(), List.t()) :: map()
-  def create_balance(groups, team_count, opts \\ []) do
+  @spec create_balance([BT.player_group()], non_neg_integer, list) :: map
+  def create_balance([], _team_count, _opts) do
+    %{
+      logs: [],
+      time_taken: 0,
+      captains: %{},
+      deviation: 0,
+      ratings: %{},
+      team_groups: %{},
+      team_players: %{},
+      team_sizes: %{},
+      means: %{},
+      stdevs: %{}
+    }
+  end
+
+  def create_balance(groups, team_count, opts) do
     start_time = System.system_time(:microsecond)
 
     # We perform all our group calculations here and assign each group
@@ -57,22 +104,39 @@ defmodule Teiserver.Battle.BalanceLib do
         }
       end)
 
-    original_parties = get_parties(expanded_groups)
+    # Now we pass this to the algorithm and it does the rest!
+    balance_result =
+      case algorithm_modules()[opts[:algorithm] || "loser_picks"] do
+        nil ->
+          raise "No balance module by the name of '#{opts[:algorithm] || "loser_picks"}'"
 
-    # raise "Call"
-    {team_groups, logs} =
-      case opts[:algorithm] || :loser_picks do
-        :loser_picks ->
-          LoserPicksAlgorithm.loser_picks(expanded_groups, team_count, opts)
-        :cheeky_switcher ->
-          CheekySwitcherAlgorithm.cheeky_switcher(expanded_groups, team_count, opts)
-        :cheeky_switcher_rating ->
-          CheekySwitcherAlgorithm.cheeky_switcher_rating(expanded_groups, team_count, opts)
-        :cheeky_switcher_smart ->
-          CheekySwitcherSmartAlgorithm.cheeky_switcher(expanded_groups, team_count, opts)
-        :brute_force ->
-          BruteForceAlgorithm.brute_force_dont_use_in_production_for_the_love_of_bar(expanded_groups, team_count, opts)
+        m ->
+          m.perform(expanded_groups, team_count, opts)
       end
+
+    # Now expand the results and calculate stats
+    balance_result
+    |> expand_balance_result()
+    |> calculate_balance_stats
+    |> cleanup_result
+    |> Map.put(:time_taken, System.system_time(:microsecond) - start_time)
+  end
+
+  # Removes various keys we don't care about
+  defp cleanup_result(result) do
+    Map.take(
+      result,
+      ~w(team_groups team_players ratings captains team_sizes deviation means stdevs logs)a
+    )
+  end
+
+  # Take the balance result and add some extra fields to make using it easier
+  defp expand_balance_result(balance_result) do
+    team_groups =
+      balance_result.teams
+      |> Map.new(fn {team_id, groups} ->
+        {team_id, Enum.reverse(groups)}
+      end)
 
     team_players =
       team_groups
@@ -85,36 +149,212 @@ defmodule Teiserver.Battle.BalanceLib do
         {team, players}
       end)
 
-    parties_preserved =
-      original_parties
-      |> Enum.filter(fn party ->
-        team_groups
-        |> Map.values()
-        |> Enum.any?(fn team_groups ->
-          Enum.any?(team_groups, fn group ->
-            Enum.all?(party.members, fn m -> Enum.member?(group.members, m) end)
-          end)
-        end)
+    Map.merge(balance_result, %{
+      team_groups: team_groups,
+      team_players: team_players
+    })
+  end
+
+  @doc """
+  We return a list of groups, a list of solo players and logs generated in the process
+  the purpose of this function is to go through the groups, work out which ones we can keep as
+  groups and with the ones we can't, break them up and add them back into the pool of solo
+  players for other groups
+  """
+  @spec matchup_groups([BT.expanded_group()], [BT.expanded_group()], list()) ::
+          {[BT.expanded_group()], [BT.expanded_group()], [String.t()]}
+  def matchup_groups([], solo_players, _opts), do: {[], solo_players, []}
+
+  def matchup_groups(groups, solo_players, opts) do
+    # First we want to re-sort these groups, we want to have the ones with the highest standard
+    # deviation looked at first, they are the least likely to be able to be matched but most likely to
+    # help match others
+    groups =
+      groups
+      |> Enum.sort_by(
+        fn group ->
+          {group.count, Statistics.stdev(group.ratings)}
+        end,
+        &<=/2
+      )
+
+    do_matchup_groups(groups ++ solo_players, [], [], opts)
+  end
+
+  # First argument is a list of groups (size 1 included) that need to be paired
+  # the second argument is a list of the logs built up by the function
+  # thirdly is a list of already paired up groups (so can't be paired up further)
+  # fourth is the opts list
+  # the function returns a tuple of
+  # 1: paired groups
+  # 2: non-paired groups
+  # 3: logs
+  @spec do_matchup_groups([BT.expanded_group()], [String.t()], [BT.expanded_group()], list()) ::
+          {[BT.expanded_group()], [BT.expanded_group()], [String.t()]}
+  # No groups, no logs
+  defp do_matchup_groups([], [], previous_paired_groups, _opts) do
+    {previous_paired_groups, [], []}
+  end
+
+  # No remaining groups but have some logs
+  defp do_matchup_groups([], logs, previous_paired_groups, _opts) do
+    {previous_paired_groups, [], logs ++ ["End of pairing"]}
+  end
+
+  # This matches when the next group is a size 1, we no longer need to pair up
+  defp do_matchup_groups(
+         [%{count: 1} | _] = remaining_players,
+         logs,
+         previous_paired_groups,
+         _opts
+       ) do
+    {previous_paired_groups, remaining_players, logs ++ ["End of pairing"]}
+  end
+
+  # Main function clause
+  defp do_matchup_groups([group | remaining_groups], logs, previous_paired_groups, opts) do
+    group_mean = Enum.sum(group.ratings) / Enum.count(group.ratings)
+    group_stddev = Statistics.stdev(group.ratings)
+
+    {_remaining_solo, found_groups} =
+      1..(opts[:team_count] - 1)
+      |> Enum.reduce({remaining_groups, []}, fn _, {groups_to_search, results} ->
+        result = find_comparable_group(group, groups_to_search, opts)
+
+        new_groups_to_search =
+          case result do
+            :no_possible_combinations ->
+              []
+
+            :no_possible_players ->
+              []
+
+            %{members: found_members} ->
+              groups_to_search
+              |> Enum.reject(fn %{members: members} ->
+                members
+                |> Enum.any?(fn userid -> Enum.member?(found_members, userid) end)
+              end)
+          end
+
+        {new_groups_to_search, [result | results]}
       end)
 
-    parties = {Enum.count(parties_preserved), Enum.count(original_parties)}
+    case hd(found_groups) do
+      :no_possible_combinations ->
+        extra_solos =
+          Enum.zip(group.members, group.ratings)
+          |> Enum.map(fn {userid, rating} ->
+            %{
+              count: 1,
+              group_rating: rating,
+              members: [userid],
+              ratings: [rating]
+            }
+          end)
 
-    time_taken = System.system_time(:microsecond) - start_time
+        names =
+          group.members
+          |> Enum.map_join(", ", fn userid -> Account.get_username_by_id(userid) || userid end)
 
-    %{
-      team_groups: team_groups,
-      team_players: team_players,
-      logs: logs,
-      time_taken: time_taken,
-      parties: parties
-    }
-    |> calculate_balance_stats
+        pairing_logs = [
+          "Unable to find a combination match for group of #{names} (stats: #{Enum.sum(group.ratings) |> round(2)}, #{group_mean |> round(2)}, #{group_stddev |> round(2)}), treating them as solo players"
+        ]
+
+        do_matchup_groups(
+          remaining_groups ++ extra_solos,
+          logs ++ pairing_logs,
+          previous_paired_groups,
+          opts
+        )
+
+      :no_possible_players ->
+        extra_solos =
+          Enum.zip(group.members, group.ratings)
+          |> Enum.map(fn {userid, rating} ->
+            %{
+              count: 1,
+              group_rating: rating,
+              members: [userid],
+              ratings: [rating]
+            }
+          end)
+
+        names =
+          group.members
+          |> Enum.map_join(", ", fn userid -> Account.get_username_by_id(userid) || userid end)
+
+        pairing_logs = [
+          "Unable to find a player match for group of #{names} (stats: #{Enum.sum(group.ratings) |> round(2)}, #{group_mean |> round(2)}, #{group_stddev |> round(2)}), treating them as solo players"
+        ]
+
+        do_matchup_groups(
+          remaining_groups ++ extra_solos,
+          logs ++ pairing_logs,
+          previous_paired_groups,
+          opts
+        )
+
+      _ ->
+        # Calculate remaining solo players
+        combined_member_ids =
+          found_groups
+          |> Enum.map(fn g -> g.members end)
+          |> List.flatten()
+
+        remaining_groups =
+          remaining_groups
+          |> Enum.reject(fn %{members: members} ->
+            members
+            |> Enum.any?(fn userid -> Enum.member?(combined_member_ids, userid) end)
+          end)
+
+        # Generate log lines, using fgroup so it doesn't clash with group used
+        # earlier
+        grouped_logs =
+          [group | found_groups]
+          |> Enum.map(fn fgroup ->
+            fgroup_name =
+              fgroup.members
+              |> Enum.map_join(", ", fn userid -> Account.get_username_by_id(userid) || userid end)
+
+            fgroup_mean = Enum.sum(fgroup.ratings) / Enum.count(fgroup.ratings)
+            fgroup_stddev = Statistics.stdev(fgroup.ratings)
+
+            [
+              "> Grouped: #{fgroup_name}",
+              "--- Rating sum: #{fgroup.group_rating |> round(2)}",
+              "--- Rating Mean: #{fgroup_mean |> round(2)}",
+              "--- Rating Stddev: #{fgroup_stddev |> round(2)}"
+            ]
+          end)
+          |> List.flatten()
+
+        logs = ["Group matching" | logs]
+
+        # Now order the groups by rating so we can pick in the right order
+        found_groups =
+          [group | found_groups]
+          |> Enum.sort_by(fn fg -> fg.group_rating end, &>=/2)
+
+        do_matchup_groups(
+          remaining_groups,
+          logs ++ grouped_logs,
+          [found_groups | previous_paired_groups],
+          opts
+        )
+    end
   end
 
   @doc """
   Given a map from create_balance it will add in some stats
   """
-  @spec calculate_balance_stats(map()) :: map()
+
+  # @spec calculate_balance_stats(map()) :: map()
+  # def calculate_balance_stats(%{team_players: []}) do
+
+  # end
+
   def calculate_balance_stats(data) do
     ratings =
       data.team_groups
@@ -122,25 +362,29 @@ defmodule Teiserver.Battle.BalanceLib do
         {k, sum_group_rating(groups)}
       end)
 
-    # The first group in the list will be the highest ranked
-    # we take the captain as the first member of that group
+    # The highest rated member of each team is the "captain" by default
     captains =
-      data.team_groups
-      |> Map.new(fn {k, groups} ->
-        case groups do
-          [] ->
-            {k, nil}
+      if Map.has_key?(data, :captains) do
+        data.captains
+      else
+        data.team_players
+        |> Map.new(fn
+          {team_id, []} ->
+            {team_id, nil}
 
-          _ ->
-            captain =
-              groups
+          {team_id, players} ->
+            top_player =
+              players
+              |> Enum.map(fn userid ->
+                {ratings[userid], userid}
+              end)
+              |> Enum.sort_by(fn v -> v end, &>=/2)
               |> hd
-              |> Map.get(:members)
-              |> hd
+              |> elem(1)
 
-            {k, captain}
-        end
-      end)
+            {team_id, top_player}
+        end)
+      end
 
     team_sizes =
       data.team_players
@@ -191,7 +435,7 @@ defmodule Teiserver.Battle.BalanceLib do
   end
 
   @spec get_user_rating_value_uncertainty_pair(T.userid(), String.t() | non_neg_integer()) ::
-          {rating_value(), number()}
+          {BT.rating_value(), number()}
   def get_user_rating_value_uncertainty_pair(userid, rating_type_id)
       when is_integer(rating_type_id) do
     rating = Account.get_rating(userid, rating_type_id) || default_rating()
@@ -210,7 +454,7 @@ defmodule Teiserver.Battle.BalanceLib do
   @doc """
   Used to get the rating value of the user for public/reporting purposes
   """
-  @spec get_user_rating_value(T.userid(), String.t() | non_neg_integer()) :: rating_value()
+  @spec get_user_rating_value(T.userid(), String.t() | non_neg_integer()) :: BT.rating_value()
   def get_user_rating_value(userid, rating_type_id) when is_integer(rating_type_id) do
     Account.get_rating(userid, rating_type_id) |> convert_rating()
   end
@@ -227,7 +471,7 @@ defmodule Teiserver.Battle.BalanceLib do
   different from public/reporting
   """
   @spec get_user_balance_rating_value(T.userid(), String.t() | non_neg_integer()) ::
-          rating_value()
+          BT.rating_value()
   def get_user_balance_rating_value(userid, rating_type_id) when is_integer(rating_type_id) do
     real_rating = get_user_rating_value(userid, rating_type_id)
 
@@ -247,7 +491,7 @@ defmodule Teiserver.Battle.BalanceLib do
   @doc """
   Given a Rating object or nil, return a value representing the rating to be used
   """
-  @spec convert_rating(map() | nil) :: rating_value()
+  @spec convert_rating(map() | nil) :: BT.rating_value()
   def convert_rating(nil) do
     default_rating() |> convert_rating()
   end
@@ -296,6 +540,30 @@ defmodule Teiserver.Battle.BalanceLib do
     end
   end
 
+  @doc """
+  Given a list of groups, return the combined number of members
+  """
+  @spec sum_group_membership_size([BT.expanded_group()]) :: non_neg_integer()
+  def sum_group_membership_size([]), do: 0
+
+  def sum_group_membership_size(groups) do
+    groups
+    |> Enum.map(fn %{count: count} -> count end)
+    |> Enum.sum()
+  end
+
+  @doc """
+  Given a list of groups, return the combined rating (summed)
+  """
+  @spec sum_group_rating([BT.expanded_group()]) :: non_neg_integer()
+  def sum_group_rating([]), do: 0
+
+  def sum_group_rating(groups) do
+    groups
+    |> Enum.map(fn %{group_rating: group_rating} -> group_rating end)
+    |> Enum.sum()
+  end
+
   @spec calculate_leaderboard_rating(number(), number()) :: number()
   def calculate_leaderboard_rating(skill, uncertainty) do
     max(skill - 3 * uncertainty, 0)
@@ -326,6 +594,121 @@ defmodule Teiserver.Battle.BalanceLib do
     _method3 = sum + Enum.max(ratings)
     _method4 = sum + mean
     _highest_rank = Enum.max(ratings) * count
+
     sum
+  end
+
+  # Stage one, filter out players notably better/worse than the party
+  @spec find_comparable_group(BT.expanded_group(), [BT.expanded_group()], list()) ::
+          :no_possible_players | :no_possible_combinations | BT.expanded_group()
+  defp find_comparable_group(group, solo_players, opts) do
+    rating_lower_bound =
+      Enum.min(group.ratings) - (opts[:rating_lower_boundary] || @rating_lower_boundary)
+
+    rating_upper_bound =
+      Enum.max(group.ratings) + (opts[:rating_upper_boundary] || @rating_upper_boundary)
+
+    possible_players =
+      solo_players
+      |> Enum.filter(fn solo ->
+        solo.group_rating > rating_lower_bound or solo.group_rating < rating_upper_bound
+      end)
+
+    if Enum.count(possible_players) < group.count do
+      :no_possible_players
+    else
+      filter_down_possibles(group, possible_players, opts)
+    end
+  end
+
+  # Now we've trimmed our playerlist a bit lets check out the different combinations
+  @spec filter_down_possibles(BT.expanded_group(), [BT.expanded_group()], list()) ::
+          :no_possible_combinations | BT.expanded_group()
+  defp filter_down_possibles(group, possible_players, opts) do
+    group_mean = Enum.sum(group.ratings) / Enum.count(group.ratings)
+    group_stddev = Statistics.stdev(group.ratings)
+
+    sorted_possible_players =
+      possible_players
+      |> Enum.sort_by(fn g -> Enum.count(g.members) end, &>=/2)
+
+    all_combinations =
+      make_combinations(group.count, sorted_possible_players)
+
+      # Filter out bad data (parties can cause bad group sizes)
+      |> Stream.filter(fn members ->
+        total_count =
+          members
+          |> Enum.map(fn g -> g.count end)
+          |> Enum.sum()
+
+        cond do
+          total_count > group.count -> false
+          true -> true
+        end
+      end)
+
+      # This part we are getting the relevant stat info to filter on
+      |> Stream.map(fn members ->
+        member_ratings = Enum.map(members, fn %{group_rating: group_rating} -> group_rating end)
+
+        members_mean = Enum.sum(member_ratings) / group.count
+        members_stddev = Statistics.stdev(member_ratings)
+
+        mean_diff = abs(group_mean - members_mean)
+        stddev_diff = abs(group_stddev - members_stddev)
+
+        {members, mean_diff, stddev_diff}
+      end)
+
+      # We now filter on differences in mean and stddev
+      |> Stream.filter(fn {_members, mean_diff, stddev_diff} ->
+        cond do
+          mean_diff > (opts[:mean_diff_max] || @mean_diff_max) -> false
+          stddev_diff > (opts[:stddev_diff_max] || @stddev_diff_max) -> false
+          true -> true
+        end
+      end)
+
+      # Finally we sort
+      |> Enum.sort_by(
+        fn
+          {_members, mean_diff, stddev_diff} -> {mean_diff * stddev_diff, mean_diff, stddev_diff}
+        end,
+        &<=/2
+      )
+
+    case all_combinations do
+      [] ->
+        :no_possible_combinations
+
+      _ ->
+        {selected_group, _, _} = hd(all_combinations)
+
+        # Now turn a list of groups into one group
+        selected_group
+        |> Enum.reduce(%{members: [], ratings: [], count: 0, group_rating: 0}, fn solo, acc ->
+          %{
+            members: acc.members ++ solo.members,
+            ratings: acc.ratings ++ solo.ratings,
+            count: acc.count + solo.count,
+            group_rating: acc.group_rating + solo.group_rating
+          }
+        end)
+    end
+  end
+
+  # First argument is the size of each combination
+  # Second is the list of items to make a combination from
+  @spec make_combinations(integer(), list) :: [list]
+  defp make_combinations(0, _), do: [[]]
+  defp make_combinations(_, []), do: []
+
+  defp make_combinations(n, [x | xs]) do
+    if n < 0 do
+      [[]]
+    else
+      for(y <- make_combinations(n - x.count, xs), do: [x | y]) ++ make_combinations(n, xs)
+    end
   end
 end
